refactor(data table manager): refactor data table types location
