+++ packages/components/data-table-manager/src/data-table-manager-provider/index.ts
@@ -2,7 +2,5 @@ export {
   DataTableManagerProvider,
   useDataTableManagerContext,
 } from './data-table-manager-provider';
-export type {
-  TDataTableManagerContext,
-  TDataTableManagerColumns,
-} from './data-table-manager-provider';
+
+export type { TDataTableManagerContext } from './data-table-manager-provider';
